refactor: 불필요한 API 컨트롤러 제거, 사용중인 API 컨트롤러 이름 변경
