--- src/main/java/wooteco/subway/auth/ui/CustomAuthController.java
@@ -1,26 +1,0 @@
-package wooteco.subway.auth.ui;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-import wooteco.subway.auth.application.AuthService;
-import wooteco.subway.auth.dto.TokenRequest;
-import wooteco.subway.auth.dto.TokenResponse;
-
-@RestController
-public class CustomAuthController {
-
-    private final AuthService authService;
-
-    public CustomAuthController(AuthService authService) {
-        this.authService = authService;
-    }
-
-    @PostMapping("/api/login/token")
-    public ResponseEntity<TokenResponse> login(@RequestBody TokenRequest request) {
-        TokenResponse token = authService.login(request);
-        return ResponseEntity.ok()
-            .body(token);
-    }
-}
