--- src/main/java/wooteco/subway/member/ui/CustomMemberController.java
@@ -1,61 +1,0 @@
-package wooteco.subway.member.ui;
-
-import java.net.URI;
-import javax.validation.Valid;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import wooteco.subway.auth.domain.AuthenticationPrincipal;
-import wooteco.subway.member.application.MemberService;
-import wooteco.subway.member.domain.LoginMember;
-import wooteco.subway.member.dto.MemberRequest;
-import wooteco.subway.member.dto.MemberResponse;
-import wooteco.subway.member.dto.ValidateEmailRequest;
-
-@RestController
-@RequestMapping("/api/members")
-public class CustomMemberController {
-
-    private final MemberService memberService;
-
-    public CustomMemberController(MemberService memberService) {
-        this.memberService = memberService;
-    }
-
-    @PostMapping
-    public ResponseEntity createMember(@Valid @RequestBody MemberRequest request) {
-        MemberResponse member = memberService.createMember(request);
-        return ResponseEntity.created(URI.create("/members/" + member.getId())).build();
-    }
-
-    @GetMapping("/me")
-    public ResponseEntity<MemberResponse> findMemberOfMine(@AuthenticationPrincipal LoginMember loginMember) {
-        MemberResponse member = memberService.findMember(loginMember);
-        return ResponseEntity.ok().body(member);
-    }
-
-    @GetMapping("/check-validation")
-    public ResponseEntity<Void> validDuplicateEmail(@Valid @ModelAttribute ValidateEmailRequest validateEmailRequest) {
-        memberService.validateDuplicateEmail(validateEmailRequest.getEmail());
-        return ResponseEntity.ok().build();
-    }
-
-    @PutMapping("/me")
-    public ResponseEntity<MemberResponse> updateMemberOfMine(@AuthenticationPrincipal LoginMember loginMember,
-        @Valid @RequestBody MemberRequest param) {
-        memberService.updateMember(loginMember, param);
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping("/me")
-    public ResponseEntity<MemberResponse> deleteMemberOfMine(@AuthenticationPrincipal LoginMember loginMember) {
-        memberService.deleteMember(loginMember);
-        return ResponseEntity.noContent().build();
-    }
-}
