--- src/main/java/wooteco/subway/path/ui/CustomPathController.java
@@ -1,24 +1,0 @@
-package wooteco.subway.path.ui;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-import wooteco.subway.path.application.PathService;
-import wooteco.subway.path.dto.PathResponse;
-
-@RestController
-@CrossOrigin(origins = "*", allowedHeaders = "*")
-public class CustomPathController {
-    private PathService pathService;
-
-    public CustomPathController(PathService pathService) {
-        this.pathService = pathService;
-    }
-
-    @GetMapping("/api/paths")
-    public ResponseEntity<PathResponse> findPath(@RequestParam Long source, @RequestParam Long target) {
-        return ResponseEntity.ok(pathService.findPath(source, target));
-    }
-}
