+++ src/test/java/wooteco/subway/member/MemberAcceptanceTest.java
@@ -48,7 +48,7 @@ public class MemberAcceptanceTest extends AcceptanceTest {
                 .given().log().all()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .body(memberRequest)
-                .when().post("/members")
+                .when().post("/api/members")
                 .then().log().all()
                 .extract();
     }
@@ -58,7 +58,7 @@ public class MemberAcceptanceTest extends AcceptanceTest {
                 .given().log().all()
                 .auth().oauth2(tokenResponse.getAccessToken())
                 .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when().get("/members/me")
+                .when().get("/api/members/me")
                 .then().log().all()
                 .statusCode(HttpStatus.OK.value())
                 .extract();
@@ -72,7 +72,7 @@ public class MemberAcceptanceTest extends AcceptanceTest {
                 .auth().oauth2(tokenResponse.getAccessToken())
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .body(memberRequest)
-                .when().put("/members/me")
+                .when().put("/api/members/me")
                 .then().log().all()
                 .extract();
     }
@@ -81,7 +81,7 @@ public class MemberAcceptanceTest extends AcceptanceTest {
         return RestAssured
                 .given().log().all()
                 .auth().oauth2(tokenResponse.getAccessToken())
-                .when().delete("/members/me")
+                .when().delete("/api/members/me")
                 .then().log().all()
                 .extract();
     }
