+++ src/test/java/wooteco/subway/station/StationAcceptanceTest.java
@@ -8,6 +8,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import wooteco.subway.AcceptanceTest;
+import wooteco.subway.auth.dto.TokenRequest;
+import wooteco.subway.member.dto.MemberRequest;
 import wooteco.subway.station.dto.StationRequest;
 import wooteco.subway.station.dto.StationResponse;
 
@@ -19,6 +21,11 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("지하철역 관련 기능")
 public class StationAcceptanceTest extends AcceptanceTest {
+
+    private static final String ACCESS_TOKEN = "accessToken";
+    private static final String EMAIL = "email@email.com";
+    private static final String PASSWORD = "password";
+
     private static final String 강남역 = "강남역";
     private static final String 역삼역 = "역삼역";
 
@@ -26,7 +33,8 @@ public class StationAcceptanceTest extends AcceptanceTest {
     @Test
     void createStation() {
         // when
-        ExtractableResponse<Response> response = 지하철역_생성_요청(강남역);
+        회원_생성을_요청(EMAIL, PASSWORD, 10);
+        ExtractableResponse<Response> response = 토큰과_함께_지하철역_생성_요청(강남역);
 
         // then
         지하철역_생성됨(response);
@@ -36,10 +44,11 @@ public class StationAcceptanceTest extends AcceptanceTest {
     @Test
     void createStationWithDuplicateName() {
         //given
+        회원_생성을_요청(EMAIL, PASSWORD, 10);
         지하철역_등록되어_있음(강남역);
 
         // when
-        ExtractableResponse<Response> response = 지하철역_생성_요청(강남역);
+        ExtractableResponse<Response> response = 토큰과_함께_지하철역_생성_요청(강남역);
 
         // then
         지하철역_생성_실패됨(response);
@@ -49,11 +58,12 @@ public class StationAcceptanceTest extends AcceptanceTest {
     @Test
     void getStations() {
         // given
+        회원_생성을_요청(EMAIL, PASSWORD, 10);
         StationResponse stationResponse1 = 지하철역_등록되어_있음(강남역);
         StationResponse stationResponse2 = 지하철역_등록되어_있음(역삼역);
 
         // when
-        ExtractableResponse<Response> response = 지하철역_목록_조회_요청();
+        ExtractableResponse<Response> response = 토큰과_함께_지하철역_목록_조회_요청();
 
         // then
         지하철역_목록_응답됨(response);
@@ -64,43 +74,52 @@ public class StationAcceptanceTest extends AcceptanceTest {
     @Test
     void deleteStation() {
         // given
+        회원_생성을_요청(EMAIL, PASSWORD, 10);
         StationResponse stationResponse = 지하철역_등록되어_있음(강남역);
 
         // when
-        ExtractableResponse<Response> response = 지하철역_제거_요청(stationResponse);
+        ExtractableResponse<Response> response = 토큰과_함께_지하철역_제거_요청(stationResponse);
 
         // then
         지하철역_삭제됨(response);
     }
 
     public static StationResponse 지하철역_등록되어_있음(String name) {
-        return 지하철역_생성_요청(name).as(StationResponse.class);
+        return 토큰과_함께_지하철역_생성_요청(name).as(StationResponse.class);
     }
 
-    public static ExtractableResponse<Response> 지하철역_생성_요청(String name) {
+    public static ExtractableResponse<Response> 토큰과_함께_지하철역_생성_요청(String name) {
+        String accessToken = 로그인_후_토큰_발급(new TokenRequest(EMAIL, PASSWORD)).jsonPath().get(ACCESS_TOKEN);
         StationRequest stationRequest = new StationRequest(name);
 
         return RestAssured
                 .given().log().all()
                 .body(stationRequest)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().post("/stations")
+                .auth().oauth2(accessToken)
+                .when().post("/api/stations")
                 .then().log().all()
                 .extract();
     }
 
-    public static ExtractableResponse<Response> 지하철역_목록_조회_요청() {
+    public static ExtractableResponse<Response> 토큰과_함께_지하철역_목록_조회_요청() {
+        String accessToken = 로그인_후_토큰_발급(new TokenRequest(EMAIL, PASSWORD)).jsonPath().get(ACCESS_TOKEN);
+
         return RestAssured
                 .given().log().all()
-                .when().get("/stations")
+                .auth().oauth2(accessToken)
+                .when().get("/api/stations")
                 .then().log().all()
                 .extract();
     }
 
-    public static ExtractableResponse<Response> 지하철역_제거_요청(StationResponse stationResponse) {
+    public static ExtractableResponse<Response> 토큰과_함께_지하철역_제거_요청(StationResponse stationResponse) {
+        String accessToken = 로그인_후_토큰_발급(new TokenRequest(EMAIL, PASSWORD)).jsonPath().get(ACCESS_TOKEN);
+
         return RestAssured
                 .given().log().all()
-                .when().delete("/stations/" + stationResponse.getId())
+                .auth().oauth2(accessToken)
+                .when().delete("/api/stations/" + stationResponse.getId())
                 .then().log().all()
                 .extract();
     }
@@ -132,5 +151,26 @@ public class StationAcceptanceTest extends AcceptanceTest {
                 .collect(Collectors.toList());
 
         assertThat(resultLineIds).containsAll(expectedLineIds);
+    }
+
+    public static ExtractableResponse<Response> 회원_생성을_요청(String email, String password, Integer age) {
+        MemberRequest memberRequest = new MemberRequest(email, password, age);
+
+        return RestAssured
+            .given().log().all()
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .body(memberRequest)
+            .when().post("/api/members")
+            .then().log().all()
+            .extract();
+    }
+
+    public static ExtractableResponse<Response> 로그인_후_토큰_발급(TokenRequest tokenRequest) {
+        return RestAssured.given()
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .body(tokenRequest)
+            .when()
+            .post("/api/login/token")
+            .then().extract();
     }
 }
